Factoring TableDistribution out into its own class.
